refacto translations page
