--- app/routes/orgs.$orgSlug.projects.$projectSlug.translations.tsx
@@ -1,405 +1,0 @@
-import {
-  Heading,
-  VStack,
-  Button,
-  IconButton,
-  Input,
-  Box,
-  ButtonGroup,
-  Center,
-  Table,
-  Text,
-  HStack,
-  Badge,
-  LinkBox,
-  LinkOverlay,
-  Progress,
-  Pagination,
-} from "@chakra-ui/react";
-import { useTranslation } from "react-i18next";
-import { Link, Form, useOutletContext, redirect } from "react-router";
-import {
-  LuChevronLeft,
-  LuChevronRight,
-  LuPlus,
-  LuPencil,
-  LuCopy,
-} from "react-icons/lu";
-import type { Route } from "./+types/orgs.$orgSlug.projects.$projectSlug.translations";
-import { requireUser } from "~/lib/session.server";
-import { requireOrganizationMembership } from "~/lib/organizations.server";
-import { getProjectBySlug } from "~/lib/projects.server";
-import {
-  getTranslationKeys,
-  duplicateTranslationKey,
-} from "~/lib/translation-keys.server";
-import { TextHighlight } from "../lib/highlight";
-import { isSearchTranlation } from "~/lib/translation-helper";
-
-const LIMIT = 50;
-
-type ContextType = {
-  organization: { id: string; slug: string; name: string };
-  project: { id: string; slug: string; name: string };
-  languages: Array<{ id: string; locale: string; isDefault: boolean }>;
-};
-
-export async function loader({ request, params }: Route.LoaderArgs) {
-  const user = await requireUser(request);
-  const organization = await requireOrganizationMembership(
-    user,
-    params.orgSlug,
-  );
-
-  const project = await getProjectBySlug(organization.id, params.projectSlug);
-
-  if (!project) {
-    throw new Response("Project not found", { status: 404 });
-  }
-
-  const url = new URL(request.url);
-  const search = url.searchParams.get("search") || undefined;
-  const page = parseInt(url.searchParams.get("page") || "1", 10);
-  const offset = (page - 1) * LIMIT;
-
-  const keys = await getTranslationKeys(project.id, {
-    search,
-    limit: LIMIT,
-    offset,
-  });
-
-  return { keys, search, page };
-}
-
-export async function action({ request, params }: Route.ActionArgs) {
-  const user = await requireUser(request);
-  const organization = await requireOrganizationMembership(
-    user,
-    params.orgSlug,
-  );
-
-  const project = await getProjectBySlug(organization.id, params.projectSlug);
-
-  if (!project) {
-    throw new Response("Project not found", { status: 404 });
-  }
-
-  const formData = await request.formData();
-  const action = formData.get("_action");
-
-  if (action === "duplicate") {
-    const keyId = formData.get("keyId");
-
-    if (!keyId || typeof keyId !== "string") {
-      throw new Response("Key ID is required", { status: 400 });
-    }
-
-    const parsedKeyId = parseInt(keyId, 10);
-
-    if (isNaN(parsedKeyId)) {
-      throw new Response("Invalid Key ID", { status: 400 });
-    }
-
-    const newKeyId = await duplicateTranslationKey(parsedKeyId);
-
-    return redirect(
-      `/orgs/${params.orgSlug}/projects/${params.projectSlug}/keys/${newKeyId}`,
-    );
-  }
-
-  return { error: "Action inconnue" };
-}
-
-export default function ProjectTranslations({
-  loaderData,
-}: Route.ComponentProps) {
-  const { t } = useTranslation();
-  const {
-    keys: { data, count },
-    search,
-    page,
-  } = loaderData;
-  const { organization, project, languages } = useOutletContext<ContextType>();
-
-  const totalLanguages = languages.length;
-
-  // Build redirect URL with current search params
-  const currentUrl = `/orgs/${organization.slug}/projects/${project.slug}/translations${search ? `?search=${encodeURIComponent(search)}` : ""}`;
-
-  return (
-    <VStack gap={6} align="stretch">
-      <HStack justify="space-between">
-        <Box>
-          <Heading as="h2" size="lg">
-            {t("translations.title")}
-          </Heading>
-          <Text color="gray.600" mt={2}>
-            {t("translations.count", { count })}
-          </Text>
-        </Box>
-        {languages.length > 0 && (
-          <Button asChild colorPalette="brand">
-            <Link
-              to={`/orgs/${organization.slug}/projects/${project.slug}/keys/new`}
-            >
-              <LuPlus /> Nouvelle clé {t("translations.newKey")}
-            </Link>
-          </Button>
-        )}
-      </HStack>
-
-      <Form method="get">
-        <HStack>
-          <Input
-            name="search"
-            placeholder={t("translations.searchPlaceholder")}
-            defaultValue={search}
-          />
-          <Button type="submit" colorPalette="brand">
-            {t("translations.search")}
-          </Button>
-          {search && (
-            <Button asChild variant="outline">
-              <Link
-                to={`/orgs/${organization.slug}/projects/${project.slug}/translations`}
-              >
-                {t("translations.clear")}
-              </Link>
-            </Button>
-          )}
-        </HStack>
-      </Form>
-
-      {languages.length === 0 ? (
-        <Box p={10} textAlign="center" borderWidth={1} borderRadius="lg">
-          <Text color="gray.600" mb={4}>
-            {t("translations.noLanguages")}
-          </Text>
-          <Button asChild colorPalette="brand">
-            <Link
-              to={`/orgs/${organization.slug}/projects/${project.slug}/settings`}
-            >
-              {t("translations.manageLanguages")}
-            </Link>
-          </Button>
-        </Box>
-      ) : data.length === 0 ? (
-        <Box p={8} textAlign="center" bg="gray.50" borderRadius="md">
-          <Text color="gray.600">
-            {search
-              ? t("translations.noResultsForSearch")
-              : t("translations.noKeysEmpty")}
-          </Text>
-        </Box>
-      ) : (
-        <>
-          <Table.Root variant="outline">
-            <Table.Header>
-              <Table.Row>
-                <Table.ColumnHeader>
-                  {t("translations.table.keyName")}
-                </Table.ColumnHeader>
-                <Table.ColumnHeader w="300px">
-                  {t("translations.table.translations")}
-                </Table.ColumnHeader>
-                <Table.ColumnHeader w="280px">
-                  {t("translations.table.actions")}
-                </Table.ColumnHeader>
-              </Table.Row>
-            </Table.Header>
-            <Table.Body>
-              {data.map((key) => {
-                const translatedCount = key.translatedLocales.length;
-                const progressPercent =
-                  totalLanguages > 0
-                    ? (translatedCount / totalLanguages) * 100
-                    : 0;
-
-                return (
-                  <Table.Row key={key.id}>
-                    <Table.Cell>
-                      <LinkBox>
-                        <VStack align="stretch" gap={1}>
-                          <LinkOverlay asChild>
-                            <Link
-                              to={`/orgs/${organization.slug}/projects/${project.slug}/keys/${key.id}?redirect=${encodeURIComponent(currentUrl)}`}
-                            >
-                              <Text
-                                fontFamily="mono"
-                                fontSize="sm"
-                                fontWeight="medium"
-                              >
-                                <TextHighlight
-                                  text={key.keyName}
-                                  query={search}
-                                />
-                              </Text>
-                            </Link>
-                          </LinkOverlay>
-                          {key.description && (
-                            <Text fontSize="xs" color="gray.400">
-                              <TextHighlight
-                                text={key.description}
-                                query={search}
-                              />
-                            </Text>
-                          )}
-                          {isSearchTranlation(key) && (
-                            <>
-                              {key.matchType === "translation" &&
-                                key.translationLocale && (
-                                  <HStack gap={2} mt={1}>
-                                    <Badge colorScheme="purple" size="sm">
-                                      {t("translations.badgeTranslation")}
-                                    </Badge>
-                                    <Badge colorPalette="brand" size="sm">
-                                      {key.translationLocale.toUpperCase()}
-                                    </Badge>
-                                  </HStack>
-                                )}
-                              {key.matchType === "key" && (
-                                <Badge colorScheme="purple" size="sm" mt={1}>
-                                  {t("translations.badgeKey")}
-                                </Badge>
-                              )}
-                              {key.matchType === "translation" &&
-                                key.translationValue && (
-                                  <Text fontSize="xs" color="gray.600" mt={1}>
-                                    <TextHighlight
-                                      text={key.translationValue}
-                                      query={search}
-                                    />
-                                  </Text>
-                                )}
-                              {key.matchType !== "translation" &&
-                                key.defaultTranslation && (
-                                  <Text fontSize="s" color="gray.600">
-                                    {key.defaultTranslation}
-                                  </Text>
-                                )}
-                            </>
-                          )}
-                        </VStack>
-                      </LinkBox>
-                    </Table.Cell>
-                    <Table.Cell>
-                      <VStack align="stretch" gap={2}>
-                        <HStack justify="space-between" fontSize="sm">
-                          <Text color="gray.600">
-                            {translatedCount}/{totalLanguages}
-                          </Text>
-                          <Text color="gray.600">
-                            {Math.round(progressPercent)}%
-                          </Text>
-                        </HStack>
-                        <Progress.Root
-                          value={progressPercent}
-                          size="sm"
-                          colorPalette="brand"
-                        >
-                          <Progress.Track>
-                            <Progress.Range />
-                          </Progress.Track>
-                        </Progress.Root>
-                        {key.translatedLocales.length > 0 && (
-                          <HStack gap={1} flexWrap="wrap">
-                            {key.translatedLocales.map((locale) => (
-                              <Badge
-                                key={locale}
-                                size="sm"
-                                colorPalette="brand"
-                              >
-                                {locale.toUpperCase()}
-                              </Badge>
-                            ))}
-                          </HStack>
-                        )}
-                      </VStack>
-                    </Table.Cell>
-                    <Table.Cell>
-                      <HStack gap={2}>
-                        <Button asChild size="sm" colorPalette="brand">
-                          <Link
-                            to={`/orgs/${organization.slug}/projects/${project.slug}/keys/${key.id}?redirect=${encodeURIComponent(currentUrl)}`}
-                          >
-                            <LuPencil /> {t("translations.edit")}
-                          </Link>
-                        </Button>
-                        <Form method="post">
-                          <input
-                            type="hidden"
-                            name="_action"
-                            value="duplicate"
-                          />
-                          <input type="hidden" name="keyId" value={key.id} />
-                          <Button type="submit" size="sm" variant="outline">
-                            <LuCopy /> {t("translations.duplicate")}
-                          </Button>
-                        </Form>
-                      </HStack>
-                    </Table.Cell>
-                  </Table.Row>
-                );
-              })}
-            </Table.Body>
-          </Table.Root>
-
-          <Center>
-            <Pagination.Root count={count} pageSize={LIMIT} defaultPage={1}>
-              <ButtonGroup variant="outline">
-                <Pagination.PrevTrigger asChild>
-                  <IconButton asChild>
-                    <Link
-                      to={`/orgs/${organization.slug}/projects/${project.slug}/translations?${new URLSearchParams(
-                        {
-                          ...(search && { search }),
-                          page: String(page - 1),
-                        },
-                      )}`}
-                    >
-                      <LuChevronLeft />
-                    </Link>
-                  </IconButton>
-                </Pagination.PrevTrigger>
-
-                <Pagination.Items
-                  render={(page) => (
-                    <IconButton
-                      asChild
-                      variant={{ base: "ghost", _selected: "outline" }}
-                    >
-                      <Link
-                        to={`/orgs/${organization.slug}/projects/${project.slug}/translations?${new URLSearchParams(
-                          {
-                            ...(search && { search }),
-                            page: String(page.value),
-                          },
-                        )}`}
-                      >
-                        {page.value}
-                      </Link>
-                    </IconButton>
-                  )}
-                />
-                <Pagination.NextTrigger asChild>
-                  <IconButton asChild>
-                    <Link
-                      to={`/orgs/${organization.slug}/projects/${project.slug}/translations?${new URLSearchParams(
-                        {
-                          ...(search && { search }),
-                          page: String(page + 1),
-                        },
-                      )}`}
-                    >
-                      <LuChevronRight />
-                    </Link>
-                  </IconButton>
-                </Pagination.NextTrigger>
-              </ButtonGroup>
-            </Pagination.Root>
-          </Center>
-        </>
-      )}
-    </VStack>
-  );
-}
