+++ app/routes/orgs.$orgSlug.projects.$projectSlug.translations/index.tsx
@@ -1,0 +1,179 @@
+import { Heading, VStack, Button, Box, Text, HStack } from "@chakra-ui/react";
+import { Link, useOutletContext, redirect } from "react-router";
+import { useTranslation } from "react-i18next";
+import { LuPlus } from "react-icons/lu";
+import type { Route } from "./+types/index";
+import { requireUser } from "~/lib/session.server";
+import { requireOrganizationMembership } from "~/lib/organizations.server";
+import { getProjectBySlug } from "~/lib/projects.server";
+import {
+  getTranslationKeys,
+  duplicateTranslationKey,
+} from "~/lib/translation-keys.server";
+import { TranslationsSearchBar } from "./TranslationsSearchBar";
+import { TranslationsTable } from "./TranslationsTable";
+import { TranslationsPagination } from "./TranslationsPagination";
+
+const LIMIT = 50;
+
+type ContextType = {
+  organization: { id: string; slug: string; name: string };
+  project: { id: string; slug: string; name: string };
+  languages: Array<{ id: string; locale: string; isDefault: boolean }>;
+};
+
+export async function loader({ request, params }: Route.LoaderArgs) {
+  const user = await requireUser(request);
+  const organization = await requireOrganizationMembership(
+    user,
+    params.orgSlug,
+  );
+
+  const project = await getProjectBySlug(organization.id, params.projectSlug);
+
+  if (!project) {
+    throw new Response("Project not found", { status: 404 });
+  }
+
+  const url = new URL(request.url);
+  const search = url.searchParams.get("search") || undefined;
+  const page = parseInt(url.searchParams.get("page") || "1", 10);
+  const offset = (page - 1) * LIMIT;
+
+  const keys = await getTranslationKeys(project.id, {
+    search,
+    limit: LIMIT,
+    offset,
+  });
+
+  return { keys, search, page };
+}
+
+export async function action({ request, params }: Route.ActionArgs) {
+  const user = await requireUser(request);
+  const organization = await requireOrganizationMembership(
+    user,
+    params.orgSlug,
+  );
+
+  const project = await getProjectBySlug(organization.id, params.projectSlug);
+
+  if (!project) {
+    throw new Response("Project not found", { status: 404 });
+  }
+
+  const formData = await request.formData();
+  const action = formData.get("_action");
+
+  if (action === "duplicate") {
+    const keyId = formData.get("keyId");
+
+    if (!keyId || typeof keyId !== "string") {
+      throw new Response("Key ID is required", { status: 400 });
+    }
+
+    const parsedKeyId = parseInt(keyId, 10);
+
+    if (isNaN(parsedKeyId)) {
+      throw new Response("Invalid Key ID", { status: 400 });
+    }
+
+    const newKeyId = await duplicateTranslationKey(parsedKeyId);
+
+    return redirect(
+      `/orgs/${params.orgSlug}/projects/${params.projectSlug}/keys/${newKeyId}`,
+    );
+  }
+
+  return { error: "Action inconnue" };
+}
+
+export default function ProjectTranslations({
+  loaderData,
+}: Route.ComponentProps) {
+  const { t } = useTranslation();
+  const {
+    keys: { data, count },
+    search,
+    page,
+  } = loaderData;
+  const { organization, project, languages } = useOutletContext<ContextType>();
+
+  const totalLanguages = languages.length;
+
+  // Build redirect URL with current search params
+  const currentUrl = `/orgs/${organization.slug}/projects/${project.slug}/translations${search ? `?search=${encodeURIComponent(search)}` : ""}`;
+
+  return (
+    <VStack gap={6} align="stretch">
+      <HStack justify="space-between">
+        <Box>
+          <Heading as="h2" size="lg">
+            {t("translations.title")}
+          </Heading>
+          <Text color="gray.600" mt={2}>
+            {t("translations.count", { count })}
+          </Text>
+        </Box>
+        {languages.length > 0 && (
+          <Button asChild colorPalette="brand">
+            <Link
+              to={`/orgs/${organization.slug}/projects/${project.slug}/keys/new`}
+            >
+              <LuPlus /> Nouvelle clé {t("translations.newKey")}
+            </Link>
+          </Button>
+        )}
+      </HStack>
+
+      <TranslationsSearchBar
+        search={search}
+        organizationSlug={organization.slug}
+        projectSlug={project.slug}
+      />
+
+      {languages.length === 0 ? (
+        <Box p={10} textAlign="center" borderWidth={1} borderRadius="lg">
+          <Text color="gray.600" mb={4}>
+            {t("translations.noLanguages")}
+          </Text>
+          <Button asChild colorPalette="brand">
+            <Link
+              to={`/orgs/${organization.slug}/projects/${project.slug}/settings`}
+            >
+              {t("translations.manageLanguages")}
+            </Link>
+          </Button>
+        </Box>
+      ) : data.length === 0 ? (
+        <Box p={8} textAlign="center" bg="gray.50" borderRadius="md">
+          <Text color="gray.600">
+            {search
+              ? t("translations.noResultsForSearch")
+              : t("translations.noKeysEmpty")}
+          </Text>
+        </Box>
+      ) : (
+        <>
+          <TranslationsTable
+            data={data}
+            search={search}
+            totalLanguages={totalLanguages}
+            organizationSlug={organization.slug}
+            projectSlug={project.slug}
+            currentUrl={currentUrl}
+          />
+
+          <TranslationsPagination
+            count={count}
+            pageSize={LIMIT}
+            currentPage={page}
+            search={search}
+            organizationSlug={organization.slug}
+            projectSlug={project.slug}
+          />
+        </>
+      )}
+    </VStack>
+  );
+}
